Suggest API/requirements.txt
fastapi
mysql-conector-python
python-dotenv
pandas
